// components/text/types.ts
enum TextTypes {
    h1 = 'h1',
    h2 = 'h2',
    h3 = 'h3',
    h4 = 'h4',
    h5 = 'h5',
    h6 = 'h6',
    b = 'b',
    i = 'i',
    em = 'em',
    mark = 'mark',
    del = 'del',
    strong = 'strong',
    ins = 'Ins',
    sub = 'sub',
    sup = 'sup',
    custom = 'custom',
}

export { TextTypes }